+++ src/ai/susi/server/api/cms/GetFileAtCommitID.java
@@ -2,6 +2,7 @@ package ai.susi.server.api.cms;
 
 import ai.susi.DAO;
 import ai.susi.json.JsonObjectWithDefault;
+import ai.susi.mind.SusiSkill;
 import ai.susi.server.*;
 import org.eclipse.jgit.lib.ObjectId;
 import org.eclipse.jgit.lib.ObjectLoader;
@@ -53,7 +54,7 @@ public class GetFileAtCommitID extends AbstractAPIHandler implements APIHandler 
         String language_name = call.get("language", "en");
         File language = new File(group, language_name);
         String skill_name = call.get("skill", "wikipedia");
-        File skill = DAO.getSkillFile(language, skill_name);
+        File skill = SusiSkill.getSkillFile(language, skill_name);
 
         String commitID  = call.get("commitID", null);
