+++ src/ai/susi/server/api/cms/GetSkillJsonService.java
@@ -21,6 +21,7 @@ package ai.susi.server.api.cms;
 
 import ai.susi.DAO;
 import ai.susi.json.JsonObjectWithDefault;
+import ai.susi.mind.SusiSkill;
 import ai.susi.server.*;
 import org.json.JSONObject;
 
@@ -69,7 +70,7 @@ public class GetSkillJsonService extends AbstractAPIHandler implements APIHandle
         String language_name = call.get("language", "en");
         File language = new File(group, language_name);
         String skill_name = call.get("skill", "wikipedia");
-        File skill = DAO.getSkillFile(language, skill_name);
+        File skill = SusiSkill.getSkillFile(language, skill_name);
 
         try {
             String content = new String(Files.readAllBytes(skill.toPath()));
