+++ src/ai/susi/server/api/cms/RateSkillService.java
@@ -22,6 +22,7 @@ package ai.susi.server.api.cms;
 import ai.susi.DAO;
 import ai.susi.json.JsonObjectWithDefault;
 import ai.susi.json.JsonTray;
+import ai.susi.mind.SusiSkill;
 import ai.susi.server.*;
 import org.json.JSONObject;
 
@@ -65,7 +66,7 @@ public class RateSkillService extends AbstractAPIHandler implements APIHandler {
         String language_name = call.get("language", "en");
         File language = new File(group, language_name);
         String skill_name = call.get("skill", null);
-        File skill = DAO.getSkillFile(language, skill_name);
+        File skill = SusiSkill.getSkillFile(language, skill_name);
         String skill_rate = call.get("rating", null);
 
         JSONObject result = new JSONObject();
