+++ src/ai/susi/server/api/cms/UndoDeleteSkillService.java
@@ -21,6 +21,7 @@ package ai.susi.server.api.cms;
 
 import ai.susi.DAO;
 import ai.susi.json.JsonObjectWithDefault;
+import ai.susi.mind.SusiSkill;
 import ai.susi.server.*;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
@@ -65,7 +66,7 @@ public class UndoDeleteSkillService  extends AbstractAPIHandler implements APIHa
         String language_name = call.get("language", "en");
         File language = new File(group, language_name);
         String skill_name = call.get("skill", null);
-        File skill = DAO.getSkillFile(language, skill_name);
+        File skill = SusiSkill.getSkillFile(language, skill_name);
         JSONObject json = new JSONObject(true);
         json.put("accepted", false);
         String path = skill.getPath();
